+++ internal/app/wwctl/overlay/show/main_test.go
@@ -46,12 +46,8 @@ nfs:
   export paths:
   - path: /home
     export options: rw,sync
-    mount options: defaults
-    mount: true
   - path: /opt
-    export options: ro,sync,no_root_squash
-    mount options: defaults
-    mount: false`)
+    export options: ro,sync,no_root_squash`)
 
 	env.WriteFile("etc/warewulf/nodes.conf",
 		`nodeprofiles: