+++ internal/pkg/node/fields_test.go
@@ -103,6 +103,9 @@ func Test_listFields(t *testing.T) {
 							},
 						},
 					},
+					Resources: map[string]Resource{
+						"resource": "resvalue",
+					},
 				},
 			},
 			fields: []string{
@@ -142,6 +145,7 @@ func Test_listFields(t *testing.T) {
 				"NetDevs[default].Tags[nettag]",
 				"Tags[tag]",
 				"PrimaryNetDev",
+				"Resources[resource]",
 			},
 		},
 		"profile": {
@@ -155,6 +159,9 @@ func Test_listFields(t *testing.T) {
 							"nettag": "netvalue",
 						},
 					},
+				},
+				Resources: map[string]Resource{
+					"resource": "resvalue",
 				},
 			},
 			fields: []string{
@@ -192,6 +199,7 @@ func Test_listFields(t *testing.T) {
 				"NetDevs[default].Tags[nettag]",
 				"Tags[tag]",
 				"PrimaryNetDev",
+				"Resources[resource]",
 			},
 		},
 	}